add a common layout for the personal and social feed pages
